# Conflicts:
#	src/backend/generic/biz-generic/src/main/kotlin/com/tencent/bkrepo/generic/artifact/GenericLocalRepository.kt

+++ src/backend/fs/boot-fs-server/src/main/kotlin/com/tencent/bkrepo/fs/server/response/IoaTicketResponse.kt
@@ -33,7 +33,7 @@ data class IoaTicketResponse(
     @JsonProperty("Ret")
     val ret: Int,
     @JsonProperty("ExpirationTimeOffset")
-    val expirationTimeOffset: Int,
+    val expirationTimeOffset: Long,
     @JsonProperty("ErrMsg")
     val errMsg: String,
 )
